requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-anthropic==0.1.4
anthropic==0.18.1
chromadb==0.4.22
PyGithub==2.2.0
python-dotenv==1.0.1
tiktoken==0.6.0
